docs(installation): fix VSCode setup instructions
Update Visual Studio Code installation instructions to provide a
step-by-step guide for configuring the Lexical extension to use the
Expert language server. This includes creating a startup script,
setting executable permissions, and updating extension settings.

The previous version was not working because the Lexical extension does
not add the `--stdio` option to the executable configures on the
`Server: Path to server script`, so one can pass a directory and it will
actually start the `start_lexical.sh` from that dir.