feat(code-intelligence): add go to definition for Elixir stdlib
Enable finding definitions for Elixir standard library modules and functions
by introducing the ElixirSource module. This improves go-to-definition
accuracy for built-in code, overcoming ElixirSense limitations.

This works for functions, modules, structs and macros. The code as it is
right now is totally generated using Claude Code. It's reviewed, but
there are some rough edges that I will take some time to fix on
subsequent commits to this branch